% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-rates.R
\name{get_rates}
\alias{get_rates}
\title{input a score value and extracts}
\usage{
get_rates(s, as_vec = TRUE)
}
\arguments{
\item{s}{score value}

\item{as_vec}{return as vector (TRUE) or dataset (FALSE). Default TRUE}
}
\description{
input a score value and extracts
}
